+++ src/net/maizegenetics/gbs/pipeline/MergeIdenticalTaxaPlugin.java
@@ -267,7 +267,7 @@ public class MergeIdenticalTaxaPlugin extends AbstractPlugin {
             Arrays.fill(alleleDepth, 0);
 
             for (int t = 0; t < taxaIndex.length; t++) {
-                byte[] myAlleledepth = a.depthForAlleles(taxaIndex[t], s);
+                int[] myAlleledepth = a.depthForAlleles(taxaIndex[t], s);
                 for (int al=0; al<myAlleledepth.length; al++)
                 {
                     alleleDepth[al] += (int)myAlleledepth[al];
